Added SQL formatting for Supporting SqlServerDatabaseType
